+++ NEWS.md
@@ -1,5 +1,9 @@
 # metR (development version)
 
+## New Features
+
+-   `FitLm()` gains a new `intercept` argument that allows you to remove the intercept term that is automatically added.
+
 ## Bug Fixes
 
 -   Fixes error in `Laplacian()` when computing the laplacian of a single variable. (#170, Thanks @pascaloettli)